requirements.txt
# News Application – Python dependencies
# Install with: pip install -r requirements.txt

Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
requests>=2.31
mysqlclient>=2.1        # MariaDB/MySQL database driver
Pillow>=10.0            # optional: for image fields
django-filter>=23       # optional: for filtering querysets in DRF
